… pinot-core/src/test/java/org/apache/pinot/core/operator/filter/BitmapCollectionTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/ExplainPlanQueriesTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/FastFilteredCountTest.java
… pinot-core/src/test/java/org/apache/pinot/queries/TextSearchQueriesTest.java
… pinot-perf/src/main/java/org/apache/pinot/perf/BenchmarkQueries.java