Implements the one dimensional case of #8410 

Before:
```
Benchmark               (_numRows)                                                                    (_query)  (_scenario)  Mode  Cnt     Score      Error  Units
BenchmarkQueries.query     1500000  SELECT COUNT(*) FROM MyTable WHERE INT_COL IN (0, 1, 2, 3, 4, 5, 7, 9, 10)   EXP(0.001)  avgt    5   575.001 ±  170.398  us/op
BenchmarkQueries.query     1500000  SELECT COUNT(*) FROM MyTable WHERE INT_COL IN (0, 1, 2, 3, 4, 5, 7, 9, 10)     EXP(0.5)  avgt    5  6863.184 ±  887.486  us/op
BenchmarkQueries.query     1500000  SELECT COUNT(*) FROM MyTable WHERE INT_COL IN (0, 1, 2, 3, 4, 5, 7, 9, 10)   EXP(0.999)  avgt    5  7414.995 ± 1247.503  us/op
```
After:

```
Benchmark               (_numRows)                                                                    (_query)  (_scenario)  Mode  Cnt    Score     Error  Units
BenchmarkQueries.query     1500000  SELECT COUNT(*) FROM MyTable WHERE INT_COL IN (0, 1, 2, 3, 4, 5, 7, 9, 10)   EXP(0.001)  avgt    5  713.136 ±  41.245  us/op
BenchmarkQueries.query     1500000  SELECT COUNT(*) FROM MyTable WHERE INT_COL IN (0, 1, 2, 3, 4, 5, 7, 9, 10)     EXP(0.5)  avgt    5  816.003 ± 103.969  us/op
BenchmarkQueries.query     1500000  SELECT COUNT(*) FROM MyTable WHERE INT_COL IN (0, 1, 2, 3, 4, 5, 7, 9, 10)   EXP(0.999)  avgt    5  546.664 ± 181.613  us/op
```