@Jackie-Jiang 2022-03-25
This is great optimization. For count queries, we should avoid iterating the bitmaps to construct the `DocIdSetBlock`.
The current implementation can only optimize the `BitmapBasedFilterOperator`, but ideally we want to optimize all count queries (e.g. count on sorted index can be significantly optimized to `endDocId - startDocId`). I'd suggest adding a new method `int getNumMatchingDocs()` to `BaseFilterOperator` to force all filter operators to support it, then this optimization can be applied to all queries with `count(*)` without group-by