@richardstartin 2022-03-28
> Suggest just adding `RoaringBitmap getBitmap()` without the boolean check methods. We can add default implementation of them by just scanning the `BlockDocIdIterator`, and override the filter operators that can accelerate this process.
> 
> This way, all the single `COUNT` aggregation query can be solved with the short-circuited operator. Even if there is no special optimization for the filter (scan-based), we can still save the overhead of creating 10000 docs blocks.

This increases code complexity a lot, because it involves inlining a lot of code into operators. This will eventually improve efficiency but makes the surface area of the change much larger.

 I think the approach taken here is better because 

1. the operators aren't SPIs and can be removed at any time
2. they allow incremental improvements, e.g. the ones in this PR. Extending the mechanism to some other operators is quite easy, and once all operators return true for the check methods, they can be removed.