@richardstartin 2022-03-29
A few points about enforcing invariants and preconditions came up in this review:
* it would be nice to be able to assume SV bitmap indexes are disjoint, but they won't be in general because we will soon have SV bitmap indexes which are not disjoint (text)
* Dictionary ids have been assumed to be non-empty for a long time, but we have some operators which violate this assumption (e.g. REGEXP_LIKE) but we don't model this assumption. We can risk blowing up at runtime, or write potentially dead code, it isn't clear which at the call site.
* What can we assume about an operator's structure? e.g. should an `AndFilterOperator` worry about having zero child filters?